FROMLIST: irqchip/irq-loongarch-ir:Add Redirect irqchip support
The main function of the Redirected interrupt controller is to manage
the redirected-interrupt table, which consists of many redirected entries.
When MSI interrupts are requested, the driver creates a corresponding
redirected entry that describes the target CPU/vector number and the
operating mode of the interrupt. The redirected interrupt module has an
independent cache, and during the interrupt routing process, it will
prioritize the redirected entries that hit the cache. The driver
invalidates certain entry caches via a command queue.

Co-developed-by: Liupu Wang <wangliupu@loongson.cn>
Signed-off-by: Liupu Wang <wangliupu@loongson.cn>
Signed-off-by: Tianyang Zhang <zhangtianyang@loongson.cn>

Link: https://lore.kernel.org/loongarch/20251010022845.140234-1-zhangtianyang@loongson.cn/
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>