FROMLIST: net: stmmac: Add glue driver for Motorcomm YT6801 ethernet controller
Motorcomm YT6801 is a PCIe ethernet controller based on DWMAC4 IP. It
integrates an GbE phy, supporting WOL, VLAN tagging and various types
of offloading. It ships an on-chip eFuse for storing various vendor
configuration, including MAC address.

This patch adds basic glue code for the controller, allowing it to be
set up and transmit data at a reasonable speed. Features like WOL could
be implemented in the future.

Signed-off-by: Yao Zi <ziyao@disroot.org>

Link: https://lore.kernel.org/all/20251014164746.50696-2-ziyao@disroot.org/
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>