LOONGSON: dmaengine: loongson2-mcdma: New driver for the Loongson Multi-Channel DMA controller
This DMA controller appears in Loongson-2K0300 and Loongson-2K3000.

It is a multi-channel controller that enables data transfers from memory
to memory, device to memory, and memory to device, as well as channel
prioritization configurable through the channel configuration registers.

In addition, there are slight differences between Loongson-2K0300 and
Loongson-2K3000, such as channel register offsets and the number of
channels.

Signed-off-by: Binbin Zhou <zhoubinbin@loongson.cn>

[Mingcong Bai: Preliminary patch for platform testing on OrangePi Nova.]

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>