LOONGSON: dt-bindings: dmaengine: Add Loongson Multi-Channel DMA controller
The Loongson-2K0300/Loongson-2K3000 have built-in multi-channel DMA
controllers, which are similar except for some of the register offsets
and number of channels.

Obviously, this is quite different from the APB DMA controllers of the
Loongson-2K0500/Loongson-2K1000, which are single-channel, for example.

So try to add a new yaml file to avoid the single dt-binding file being
too cluttered.

Signed-off-by: Binbin Zhou <zhoubinbin@loongson.cn>

[Mingcong Bai: Preliminary patch for platform testing on OrangePi Nova.]

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>