dt-bindings: gnss: u-blox: use open-drain reset and safeboot in example
The RESET_N and SAFEBOOT_N pins should typically be left open and only
be connected to ground in rare cases where a device is misbehaving.

Specify GPIO_OPEN_DRAIN in the example as this is what should generally
be used.

Note that the current Linux driver implementation depends on these pins
being declared open-drain so that they are not driven while the main
supply is off.

Acked-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Johan Hovold <johan@kernel.org>