gfs2: Fix use of bio_chain
In gfs2_chain_bio(), the call to bio_chain() has its arguments swapped.
On filesystems with a fragmented journal, this will eventually result in
a freeze because the bio's ->bi_end_io() function will never be called.

This code never triggers when the journals are initialized at filesystem
create time because then, the journals will be contiguous on disk.  It
will trigger when journals are created later with gfs2_jadd, though.

Reported-by: Stephen Zhang <starzhangzsd@gmail.com>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>